api/requirements.txt
uvicorn[standard]
gunicorn
fastapi
python-jose
passlib
pydantic[email]
SQLAlchemy
motor
psycopg2
python-multipart
werkzeug
unsync
svn